Solution to Find the Length of a String

+++ Find the Length of a String@@676c039d80e1714600f67643/code.c
@@ -3,7 +3,7 @@
 int main(){
     char str[400];
     scanf("%[^\n]",str);
-    printf("%d",strlen(str));
+    printf("%d",(int)strlen(str));
 
     return 0;
 }
